mytestproj/requirements.txt
asgiref==3.2.5
BeautifulReport==0.1.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
demjson==2.2.4
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
filetype==1.0.6
httprunner==3.0.1.2
idna==2.9
Jinja2==2.11.2
jsonpath==0.82
loguru==0.4.1
MarkupSafe==1.1.1
pydantic==1.5.1
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.9.1
sqlparse==0.3.1
urllib3==1.25.8
win32-setctime==1.0.1
xlrd==1.2.0
